projects/roys/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
psycopg2-binary>=2.9.9
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-docx>=1.1.0
docxtpl>=0.18.0
httpx>=0.28.0
openpyxl>=3.1.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
python-multipart>=0.0.18
